+++ hexstrike_mcp.py
@@ -279,303 +279,17 @@ def setup_mcp_server(hexstrike_client: HexStrikeClient, compact: bool = False) -
     from mcp_tools.advanced_ctf_tools import register_advanced_ctf_tools
     register_advanced_ctf_tools(mcp, hexstrike_client, logger)
 
-    # ============================================================================
-    # BUG BOUNTY RECONNAISSANCE TOOLS (v5.0 ENHANCEMENT)
-    # ============================================================================
+    # Regular bug bounty recon tools for web endpoint discovery and parameter enumeration
+    from mcp_tools.bug_bounty_recon import register_bug_bounty_recon_tools
+    register_bug_bounty_recon_tools(mcp, hexstrike_client, logger)
 
-    @mcp.tool()
-    def hakrawler_crawl(url: str, depth: int = 2, forms: bool = True, robots: bool = True, sitemap: bool = True, wayback: bool = False, additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Hakrawler for web endpoint discovery with enhanced logging.
+    # Register system monitoring tools for server health, cache stats, and telemetry
+    from mcp_tools.system_monitoring import register_system_monitoring_tools
+    register_system_monitoring_tools(mcp, hexstrike_client, logger)
 
-        Note: Uses standard Kali Linux hakrawler (hakluke/hakrawler) with parameter mapping:
-        - url: Piped via echo to stdin (not -url flag)
-        - depth: Mapped to -d flag (not -depth)
-        - forms: Mapped to -s flag for showing sources
-        - robots/sitemap/wayback: Mapped to -subs for subdomain inclusion
-        - Always includes -u for unique URLs
-
-        Args:
-            url: Target URL to crawl
-            depth: Crawling depth (mapped to -d)
-            forms: Include forms in crawling (mapped to -s)
-            robots: Check robots.txt (mapped to -subs)
-            sitemap: Check sitemap.xml (mapped to -subs)
-            wayback: Use Wayback Machine (mapped to -subs)
-            additional_args: Additional Hakrawler arguments
-
-        Returns:
-            Web endpoint discovery results
-        """
-        data = {
-            "url": url,
-            "depth": depth,
-            "forms": forms,
-            "robots": robots,
-            "sitemap": sitemap,
-            "wayback": wayback,
-            "additional_args": additional_args
-        }
-        logger.info(f"🕷️ Starting Hakrawler crawling: {url}")
-        result = hexstrike_client.safe_post("api/tools/hakrawler", data)
-        if result.get("success"):
-            logger.info(f"✅ Hakrawler crawling completed")
-        else:
-            logger.error(f"❌ Hakrawler crawling failed")
-        return result
-
-    @mcp.tool()
-    def paramspider_discovery(domain: str, exclude: str = "", output_file: str = "", level: int = 2, additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute ParamSpider for parameter discovery with enhanced logging.
-
-        Args:
-            domain: Target domain
-            exclude: Extensions to exclude
-            output_file: Output file path
-            level: Crawling level
-            additional_args: Additional ParamSpider arguments
-
-        Returns:
-            Parameter discovery results
-        """
-        data = {
-            "domain": domain,
-            "exclude": exclude,
-            "output_file": output_file,
-            "level": level,
-            "additional_args": additional_args
-        }
-        logger.info(f"🔍 Starting ParamSpider discovery: {domain}")
-        result = hexstrike_client.safe_post("api/tools/paramspider", data)
-        if result.get("success"):
-            logger.info(f"✅ ParamSpider discovery completed")
-        else:
-            logger.error(f"❌ ParamSpider discovery failed")
-        return result
-
-    # ============================================================================
-    # SYSTEM MONITORING & TELEMETRY
-    # ============================================================================
-
-    @mcp.tool()
-    def server_health() -> Dict[str, Any]:
-        """
-        Check the health status of the HexStrike AI server.
-
-        Returns:
-            Server health information with tool availability and telemetry
-        """
-        logger.info(f"🏥 Checking HexStrike AI server health")
-        result = hexstrike_client.check_health()
-        if result.get("status") == "healthy":
-            logger.info(f"✅ Server is healthy - {result.get('total_tools_available', 0)} tools available")
-        else:
-            logger.warning(f"⚠️  Server health check returned: {result.get('status', 'unknown')}")
-        return result
-
-    @mcp.tool()
-    def get_cache_stats() -> Dict[str, Any]:
-        """
-        Get cache statistics from the HexStrike AI server.
-
-        Returns:
-            Cache performance statistics
-        """
-        logger.info(f"💾 Getting cache statistics")
-        result = hexstrike_client.safe_get("api/cache/stats")
-        if "hit_rate" in result:
-            logger.info(f"📊 Cache hit rate: {result.get('hit_rate', 'unknown')}")
-        return result
-
-    @mcp.tool()
-    def clear_cache() -> Dict[str, Any]:
-        """
-        Clear the cache on the HexStrike AI server.
-
-        Returns:
-            Cache clear operation results
-        """
-        logger.info(f"🧹 Clearing server cache")
-        result = hexstrike_client.safe_post("api/cache/clear", {})
-        if result.get("success"):
-            logger.info(f"✅ Cache cleared successfully")
-        else:
-            logger.error(f"❌ Failed to clear cache")
-        return result
-
-    @mcp.tool()
-    def get_telemetry() -> Dict[str, Any]:
-        """
-        Get system telemetry from the HexStrike AI server.
-
-        Returns:
-            System performance and usage telemetry
-        """
-        logger.info(f"📈 Getting system telemetry")
-        result = hexstrike_client.safe_get("api/telemetry")
-        if "commands_executed" in result:
-            logger.info(f"📊 Commands executed: {result.get('commands_executed', 0)}")
-        return result
-
-    # ============================================================================
-    # PROCESS MANAGEMENT TOOLS (v5.0 ENHANCEMENT)
-    # ============================================================================
-
-    @mcp.tool()
-    def list_active_processes() -> Dict[str, Any]:
-        """
-        List all active processes on the HexStrike AI server.
-
-        Returns:
-            List of active processes with their status and progress
-        """
-        logger.info("📊 Listing active processes")
-        result = hexstrike_client.safe_get("api/processes/list")
-        if result.get("success"):
-            logger.info(f"✅ Found {result.get('total_count', 0)} active processes")
-        else:
-            logger.error("❌ Failed to list processes")
-        return result
-
-    @mcp.tool()
-    def get_process_status(pid: int) -> Dict[str, Any]:
-        """
-        Get the status of a specific process.
-
-        Args:
-            pid: Process ID to check
-
-        Returns:
-            Process status information including progress and runtime
-        """
-        logger.info(f"🔍 Checking status of process {pid}")
-        result = hexstrike_client.safe_get(f"api/processes/status/{pid}")
-        if result.get("success"):
-            logger.info(f"✅ Process {pid} status retrieved")
-        else:
-            logger.error(f"❌ Process {pid} not found or error occurred")
-        return result
-
-    @mcp.tool()
-    def terminate_process(pid: int) -> Dict[str, Any]:
-        """
-        Terminate a specific running process.
-
-        Args:
-            pid: Process ID to terminate
-
-        Returns:
-            Success status of the termination operation
-        """
-        logger.info(f"🛑 Terminating process {pid}")
-        result = hexstrike_client.safe_post(f"api/processes/terminate/{pid}", {})
-        if result.get("success"):
-            logger.info(f"✅ Process {pid} terminated successfully")
-        else:
-            logger.error(f"❌ Failed to terminate process {pid}")
-        return result
-
-    @mcp.tool()
-    def pause_process(pid: int) -> Dict[str, Any]:
-        """
-        Pause a specific running process.
-
-        Args:
-            pid: Process ID to pause
-
-        Returns:
-            Success status of the pause operation
-        """
-        logger.info(f"⏸️ Pausing process {pid}")
-        result = hexstrike_client.safe_post(f"api/processes/pause/{pid}", {})
-        if result.get("success"):
-            logger.info(f"✅ Process {pid} paused successfully")
-        else:
-            logger.error(f"❌ Failed to pause process {pid}")
-        return result
-
-    @mcp.tool()
-    def resume_process(pid: int) -> Dict[str, Any]:
-        """
-        Resume a paused process.
-
-        Args:
-            pid: Process ID to resume
-
-        Returns:
-            Success status of the resume operation
-        """
-        logger.info(f"▶️ Resuming process {pid}")
-        result = hexstrike_client.safe_post(f"api/processes/resume/{pid}", {})
-        if result.get("success"):
-            logger.info(f"✅ Process {pid} resumed successfully")
-        else:
-            logger.error(f"❌ Failed to resume process {pid}")
-        return result
-
-    @mcp.tool()
-    def get_process_dashboard() -> Dict[str, Any]:
-        """
-        Get enhanced process dashboard with visual status indicators.
-
-        Returns:
-            Real-time dashboard with progress bars, system metrics, and process status
-        """
-        logger.info("📊 Getting process dashboard")
-        result = hexstrike_client.safe_get("api/processes/dashboard")
-        if result.get("success", True) and "total_processes" in result:
-            total = result.get("total_processes", 0)
-            logger.info(f"✅ Dashboard retrieved: {total} active processes")
-
-            # Log visual summary for better UX
-            if total > 0:
-                logger.info("📈 Active Processes Summary:")
-                for proc in result.get("processes", [])[:3]:  # Show first 3
-                    logger.info(f"   ├─ PID {proc['pid']}: {proc['progress_bar']} {proc['progress_percent']}")
-        else:
-            logger.error("❌ Failed to get process dashboard")
-        return result
-
-    @mcp.tool()
-    def execute_command(command: str, use_cache: bool = True) -> Dict[str, Any]:
-        """
-        Execute an arbitrary command on the HexStrike AI server with enhanced logging.
-
-        Args:
-            command: The command to execute
-            use_cache: Whether to use caching for this command
-
-        Returns:
-            Command execution results with enhanced telemetry
-        """
-        try:
-            logger.info(f"⚡ Executing command: {command}")
-            result = hexstrike_client.execute_command(command, use_cache)
-            if "error" in result:
-                logger.error(f"❌ Command failed: {result['error']}")
-                return {
-                    "success": False,
-                    "error": result["error"],
-                    "stdout": "",
-                    "stderr": f"Error executing command: {result['error']}"
-                }
-
-            if result.get("success"):
-                execution_time = result.get("execution_time", 0)
-                logger.info(f"✅ Command completed successfully in {execution_time:.2f}s")
-            else:
-                logger.warning(f"⚠️  Command completed with errors")
-
-            return result
-        except Exception as e:
-            logger.error(f"💥 Error executing command '{command}': {str(e)}")
-            return {
-                "success": False,
-                "error": str(e),
-                "stdout": "",
-                "stderr": f"Error executing command: {str(e)}"
-            }
+    # Register process management tools for advanced task handling and monitoring
+    from mcp_tools.process_management import register_process_management_tools
+    register_process_management_tools(mcp, hexstrike_client, logger)
 
     # ============================================================================
     # ADVANCED VULNERABILITY INTELLIGENCE MCP TOOLS (v6.0 ENHANCEMENT)
